backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
aiosqlite>=0.20.0
networkx>=3.4
python-socketio>=5.11.0
backboard-sdk
httpx>=0.27.0